requirements.txt
bidict==0.22.0
cachelib==0.9.0
click==8.1.3
Flask==2.2.1
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SocketIO==5.3.3
future==0.18.3
importlib-metadata==4.12.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyMySQL==1.0.3
pyserial==3.5
python-engineio==4.4.1
python-socketio==5.8.0
PyYAML==6.0
six==1.11.0
Werkzeug==2.2.1
zipp==3.8.1
